% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrvec2dl.R
\name{chrvec2dl}
\alias{chrvec2dl}
\title{Convert character vector to a double list}
\usage{
chrvec2dl(x, splt = "-")
}
\arguments{
\item{x}{a character vector to be converted.}

\item{splt}{a string pattern to make defination for spliting each layer of double list.}
}
\value{
return a character double list splited by defined pattern, the default pattern is "-".
}
\description{
Structurize a character vector to a double list. Layers in the double list will be determined
by the given pattern.
}
\examples{
#example using default dataset:

charvector <- colnames(datatest)
chrvec2dl(charvector, "-")
}
\keyword{Double.List}
