\name{datatest}
\alias{datatest}
\docType{data}
\title{
A test data structurized column names.
}
\description{
A test data with 56 different columns.
}
\usage{data("datatest")}
\details{
A test data structurized column names, with two data type "Strain" and "Stress", 7 different temperatures, 4 kinds
strain rates and one level of compression rate.
}
\examples{
datatest
}
\keyword{datatest}
