% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trvs.R
\name{trvs}
\alias{trvs}
\title{Make traversal from vector binary tree}
\usage{
trvs(x)
}
\arguments{
\item{x}{A vector binary tree.}
}
\value{
Return a traversal table from the given vector binary tree.
}
\description{
Generating a table of traversal from given vector binary tree, in order to construct correct
mapping relationships within double list, vector binary tree, array and tensor.
}
\examples{
#Make vector binary tree:
colnamevbt <- dl2vbt(chrvec2dl(colnames(datatest)))

#Construct traversal table:
trvs(colnamevbt)
}
\keyword{Trav.Table}
\keyword{Vector.Binary.Tree}
