% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trvsvisit.R
\name{trvsidxinq}
\alias{trvsidxinq}
\title{Using vector to visit the traversal table}
\usage{
trvsidxinq(trvs, inq)
}
\arguments{
\item{trvs}{The traversal table to be visited, which should be generated from the vector binary tree by the function trvs().}

\item{inq}{An integer vector to assign the coordinate corresponding to the element to be visited.}
}
\value{
Return an inquiry result with a numeric item index, a character pattern and its coordinate in form of integer vector.
}
\description{
Visit the traversal table generated from a vector binary tree through the coordinate determined by the argument \code{inq}, and return
an inquiry result containing its numeric item index, its corresponding character pattern and the coordinate.
}
\examples{
#Make traversal table:
trav <- trvs(dl2vbt(chrvec2dl(colnames(datatest))))

#Visit specific element by its coordinate:
trvsidxinq(trav,c(1,2,3,1))
}
\keyword{Trav.Inq}
\keyword{Trav.Table}
